Moved SKIV_Image to their own .h/cpp files
