+++ SKIV.vcxproj
@@ -383,6 +383,7 @@
     <ClInclude Include="include\imgui\imgui_impl_dx11.h" />
     <ClInclude Include="include\utility\droptarget.hpp" />
     <ClInclude Include="include\utility\gamepad.h" />
+    <ClInclude Include="include\utility\image.h" />
     <ClInclude Include="include\utility\plog_formatter.h" />
     <ClInclude Include="packages_misc\html_coder.hpp" />
     <ClInclude Include="packages_misc\HybridDetect.h" />
@@ -428,6 +429,7 @@
     <ClCompile Include="src\imgui\imgui_impl_win32.cpp" />
     <ClCompile Include="src\imgui\imgui_tables.cpp" />
     <ClCompile Include="src\utility\gamepad.cpp" />
+    <ClCompile Include="src\utility\image.cpp" />
     <ClCompile Include="src\utility\registry.cpp" />
     <ClCompile Include="src\utility\skif_imgui.cpp" />
     <ClCompile Include="src\utility\utility.cpp" />
