+++ SKIV.vcxproj.filters
@@ -174,6 +174,9 @@
     <ClInclude Include="packages_misc\ImGuiNotify.hpp">
       <Filter>Header Files\Packages_Misc</Filter>
     </ClInclude>
+    <ClInclude Include="include\utility\image.h">
+      <Filter>Header Files\Utility</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="src\imgui\imgui.cpp">
@@ -229,6 +232,9 @@
     </ClCompile>
     <ClCompile Include="src\imgui\imgui_impl_win32.cpp">
       <Filter>Source Files\ImGui</Filter>
+    </ClCompile>
+    <ClCompile Include="src\utility\image.cpp">
+      <Filter>Source Files\Utility</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
